Support a user-task based management to replace session based managment.
CruiseContorl used sessions to manage async operations from users and
limited it by number of active sessions.

This commit introduces the concept of UserTaskManager (replaces SessionManager).
It responsilbe for managing active tasks using User-Task-ID defined in
the HTTP request/response header.

HttpSessions can still be used to fetch the status of an async operation
but it can be overridden by adding User-Task-Id header in request header.
Every async request will return the User-Task-Id in the response headers.

A new Http request is added to report the status active tasks that are in
progress.