Bump Kafka dependency to 2.3 and replace NetworkClient inside MetadataClient with the public AdminClient interface
It is better to depend on public interfaces where possible to improve API compatibility with newer Kafka versions.
This patch refactors the MetadataClient, removing the use of the Kafka's internal NetworkClient class and the associated network.client.provider.class config.
It also changes the metadata.max.age.ms config from a Long to an Integer, as the AdminClient interface uses ints for its timeouts