Enable compiling Metrics Reporter with pre-Kafka 2.4 versions (#1240)
Contrary to Kafka 2.3, AdminClient in Kafka 2.4 implements an Admin interface. The close() method of AdminClient lives in this Admin interface. Hence, method references to this method are tied to the close method of Admin in the generated bytecode.

This patch enables compiling Cruise Control (CC) with different flavors / versions of Kafka in projects that depend on CC.