Run cruise-control tests for s390x architecture
The extenstion to the current CircleCI setup allows to run the
cruise-control tests for non-Intel architectures. As first step
linux/s390x platform is used. New platforms can be added to the
existing matrix, once there is interest for it.

The idea is to run docker container on standard Intel host via standard
CirclCI setup in hardware emulation mode via dockere --platform parameter
and do the tests inside the container, similar to how it can be executed
on real ARM, Power or Z hardware. To make sure which architecture is emulated
inside the container - `uname -a` command is executed before the actual test.

Signed-off-by: Yulia Gaponenko <yulia.gaponenko1@de.ibm.com>