Support auto adjusting both broker and cluster leader move concurrency (#2051)
This PR resolves #<2056>.
Concurrency Adjuster currently does not take effect for leadership concurrency. This issue is because cc currently adjusts per broker concurrency (the max number of partitions to elect leader in a broker). However, for leadership movement, the cluster concurrency (the max total number of partitions to elect leader in a cluster) equals broker concurrency initially, of which the value is set by num.concurrent.leader.movements. That results in increasing per broker concurrency would not increase partitions in one request because it is still bounded by the cluster concurrency.

This PR is to fix the above issue.

In this PR:
(1) support configuring broker and cluster leadership movement concurrency separately (num.concurrent.leader.movements.per.broker and num.concurrent.leader.movements)
(2) support auto adjusting both broker and cluster leader move concurrency.
(3) support parameter broker_concurrent_leader_movements in Rest API for setting the upper bound of per broker ongoing leadership movements

The leadership movement task is more bottleneck on controller, so in general, the cluster allowed concurrency is smaller than the broker allowed concurrency multiplied by the number of brokers in the cluster. It would be safer and more efficient to have both cluster and broker concurrency adjusters for leadership movements. If with broker concurrency adjustor alone, the execution of leadership movements might be unnecessarily slow due to low cluster concurrency; if with cluster concurrency adjuster alone, we do not have the ability to limit the traffic to slow brokers.