Rectify docker run command for s390x
Updated temurin image to 11-jdk so that it can be in sync with ubuntu-latest image.
Added correct way of running multiple instructions during docker run command.
Added uname command in end to confirm that build ran on s390x.
Introduced continue-on-error due to intermittent test case failures on temurin container.

Signed-off-by: Yasir Ashfaq <Yasir.Ashfaq1@ibm.com>